requirements.txt
Flask
python-dotenv
openai>=1.3.5

